requirements.txt
Django>=4.2
djangorestframework
gunicorn
psycopg2-binary
whitenoise
django-cors-headers
django-heroku
